+++ 133.cpp
@@ -32,7 +32,7 @@ bool is_factor(int p)
         for (int b = 0; b <= b_max; b++)
         {
             unsigned long int q = std::pow(2, a) * std::pow(5, b);
-            if (util::pow_mod(10, q, 9 * p) == 1)
+            if (util::pow_mod(uint64_t(10), q, uint64_t(9 * p)) == 1)
                 return true;
         }
     }
